make room for different DNS management backends
Right now we just support cloudflare API for the backend management of
the DNS records. Finalize the ddns interface to abstract cloudflare API
access and be ready for alternative DNS management backends.

Fixes: #7
